+++ pkg/identityserver/entity_access.go
@@ -39,11 +39,11 @@ var (
 	errInvalidAuthorization     = errors.DefineUnauthenticated("invalid_authorization", "invalid authorization")
 	errTokenNotFound            = errors.DefineUnauthenticated("token_not_found", "token not found")
 	errTokenExpired             = errors.DefineUnauthenticated("token_expired", "token expired")
-	errUserRejected             = errors.DefinePermissionDenied("user_rejected", "user account was rejected")
-	errUserRequested            = errors.DefinePermissionDenied("user_requested", "user account approval is pending")
-	errUserSuspended            = errors.DefinePermissionDenied("user_suspended", "user account was suspended")
-	errOAuthClientRejected      = errors.DefinePermissionDenied("oauth_client_rejected", "OAuth client was rejected")
-	errOAuthClientSuspended     = errors.DefinePermissionDenied("oauth_client_suspended", "OAuth client was suspended")
+	errUserRejected             = errors.DefinePermissionDenied("user_rejected", "user account was rejected", "description")
+	errUserRequested            = errors.DefinePermissionDenied("user_requested", "user account approval is pending", "description")
+	errUserSuspended            = errors.DefinePermissionDenied("user_suspended", "user account was suspended", "description")
+	errOAuthClientRejected      = errors.DefinePermissionDenied("oauth_client_rejected", "OAuth client was rejected", "description")
+	errOAuthClientSuspended     = errors.DefinePermissionDenied("oauth_client_suspended", "OAuth client was suspended", "description")
 	errPermissionDenied         = errors.DefinePermissionDenied("permission_denied", "unauthorized request to restricted resource")
 )
 
@@ -96,8 +96,8 @@ func (is *IdentityServer) authInfo(ctx context.Context) (info *ttnpb.AuthInfoRes
 
 	var fetch func(db *gorm.DB) error
 	res := &ttnpb.AuthInfoResponse{}
-	userFieldMask := &types.FieldMask{Paths: []string{"admin", "state", "primary_email_address_validated_at"}}
-	clientFieldMask := &types.FieldMask{Paths: []string{"state"}}
+	userFieldMask := &types.FieldMask{Paths: []string{"admin", "state", "state_description", "primary_email_address_validated_at"}}
+	clientFieldMask := &types.FieldMask{Paths: []string{"state", "state_description"}}
 	var user *ttnpb.User
 	var userRights *ttnpb.Rights
 
@@ -186,10 +186,16 @@ func (is *IdentityServer) authInfo(ctx context.Context) (info *ttnpb.AuthInfoRes
 			case ttnpb.STATE_APPROVED:
 				// Normal OAuth client.
 			case ttnpb.STATE_REJECTED:
+				if client.StateDescription != "" {
+					return errOAuthClientRejected.WithAttributes("description", client.StateDescription)
+				}
 				return errOAuthClientRejected.New()
 			case ttnpb.STATE_FLAGGED:
 				// Innocent until proven guilty.
 			case ttnpb.STATE_SUSPENDED:
+				if client.StateDescription != "" {
+					return errOAuthClientSuspended.WithAttributes("description", client.StateDescription)
+				}
 				return errOAuthClientSuspended.New()
 			default:
 				panic(fmt.Sprintf("Unhandled client state: %s", client.State.String()))
@@ -270,13 +276,21 @@ func (is *IdentityServer) authInfo(ctx context.Context) (info *ttnpb.AuthInfoRes
 		case ttnpb.STATE_REJECTED:
 			// Go to profile page, delete account.
 			restrictRights(res, ttnpb.RightsFrom(ttnpb.RIGHT_USER_INFO, ttnpb.RIGHT_USER_DELETE))
-			warning.Add(ctx, "Restricted rights after account rejection")
+			if user.StateDescription != "" {
+				warning.Add(ctx, fmt.Sprintf("Restricted rights after account rejection: %s", user.StateDescription))
+			} else {
+				warning.Add(ctx, "Restricted rights after account rejection")
+			}
 		case ttnpb.STATE_FLAGGED:
 			// Innocent until proven guilty.
 		case ttnpb.STATE_SUSPENDED:
 			// Go to profile page.
 			restrictRights(res, ttnpb.RightsFrom(ttnpb.RIGHT_USER_INFO))
-			warning.Add(ctx, "Restricted rights after account suspension")
+			if user.StateDescription != "" {
+				warning.Add(ctx, fmt.Sprintf("Restricted rights after account suspension: %s", user.StateDescription))
+			} else {
+				warning.Add(ctx, "Restricted rights after account suspension")
+			}
 		default:
 			panic(fmt.Sprintf("Unhandled user state: %s", user.State.String()))
 		}
@@ -309,10 +323,19 @@ func (is *IdentityServer) RequireAuthenticated(ctx context.Context) error {
 				// Flagged users have the same authentication presence as approved users until proven guilty.
 				return nil
 			case ttnpb.STATE_REQUESTED:
+				if user.StateDescription != "" {
+					return errUserRequested.WithAttributes("description", user.StateDescription)
+				}
 				return errUserRequested.New()
 			case ttnpb.STATE_REJECTED:
+				if user.StateDescription != "" {
+					return errUserRejected.WithAttributes("description", user.StateDescription)
+				}
 				return errUserRejected.New()
 			case ttnpb.STATE_SUSPENDED:
+				if user.StateDescription != "" {
+					return errUserSuspended.WithAttributes("description", user.StateDescription)
+				}
 				return errUserSuspended.New()
 			default:
 				panic(fmt.Sprintf("Unhandled user state: %s", user.State.String()))