+++ CHANGELOG.md
@@ -12,6 +12,8 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 - API and CLI commands for listing, searching for and restoring recently deleted applications, OAuth clients, gateways, organizations and users.
+- State Description fields for adding context to the (rejected, flagged, suspended) state of Users and OAuth clients.
+  - This requires a database schema migration (`ttn-lw-stack is-db migrate`) because of the added columns.
 
 ### Changed